requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
Flask>=2.0.0
Flask-CORS>=3.0.10
gradio>=3.0.0
